refactor(offline): promote lib/offline files into directory modules
Per review feedback: every `src/lib/offline/X.{ts,tsx}` with a companion
`X.test.{ts,tsx}` is now its own directory module, matching the
convention used throughout `src/lib/utils/{arrays,guards,objects,sets}/`.

Restructured:
- formatOfflineQueryError/
- isAppLinkAvailableOffline/
- queryPersister/
- useOfflineGate/

Unmoved (no companion file → flat per the same convention):
- registerServiceWorker.ts
- useLocalDatasetIds.ts

Test-file imports updated to the new full paths. main.tsx's
queryPersister import also updated.
