+++ src/lib/offline/formatOfflineQueryError/formatOfflineQueryError.ts
