+++ src/lib/offline/useOfflineGate/useOfflineGate.test.tsx
@@ -1,7 +1,7 @@
 import { I18nProvider } from "@lingui/react";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { i18n } from "@/i18n/i18n";
-import { useOfflineGate } from "@/lib/offline/useOfflineGate";
+import { useOfflineGate } from "@/lib/offline/useOfflineGate/useOfflineGate";
 import { renderHook } from "@/test-utils";
 
 vi.mock("@ui", () => {
